feat: Move to global utils
