+++ src/common/utils/scraping_utils.ts
@@ -1,3 +1,4 @@
+import { Page } from "playwright";
 import { ProductModel } from "../core";
 
 export function printProduct(product: ProductModel) {
@@ -20,4 +21,11 @@ export function logExecutionTime(ms: number) {
     const minutes = Math.floor(ms / 60000);
     const seconds = Math.floor((ms % 60000) / 1000);
     return `${minutes}m ${seconds}s`;
+}
+
+export async function humanLikeMouseMovement(page: Page, startX : any, startY : any, endX : any, endY : any) {
+    await page.mouse.move(startX, startY);
+    await page.waitForTimeout(100 + Math.random() * 100); // Add some randomness in the movement
+    await page.mouse.wheel(100 + Math.random(), 1000);
+    await page.mouse.move(endX, endY, { steps: 5 + Math.random() * 5 });
 }
